pyproject.toml
[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "puregym_bot"
module-root = ""

[project]
name = "puregym-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pydantic-settings[yaml]>=2.13.1",
    "python-telegram-bot[job-queue]>=22.6",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[tool.ruff]
line-length=110
